requirements.txt
Cycler
Pillow
awslambdaric
cPython
certifi
chardet
idna
kiwisolver
lxml
matplotlib
mplfinance
numpy
pandas-datareader
pandas
pyparsing
python-dateutil
python-dotenv
pytz
six
urllib3
boto3
tweepy
